% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ff.R
\name{theme_ff}
\alias{theme_ff}
\title{FUNCTION_TITLE}
\usage{
theme_ff(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{ggtheme}}
}
